requirements.txt
Flask
gunicorn 
psycopg2-binary 
Flask-SQLAlchemy 
python-dotenv 
flask_restx 
flask_cors
requests
pandas
Flask-RESTX
SQLAlchemy
firebase-admin
PyJWT